refactor(plugins): ungroup `plugins.editor`
