--- lua/deltavim/plugins/dressing/initialize.lua
@@ -1,3 +1,0 @@
-return function()
-  require("astrocore").load_plugin_with_func("dressing.nvim", vim.ui, { "input", "select" })
-end
