--- lua/deltavim/plugins/dressing/opts.lua
@@ -1,4 +1,0 @@
-return {
-  input = { default_prompt = require("astroui").get_icon "Prompt" },
-  select = { backend = { "telescope", "builtin" } },
-}
