+++ lua/deltavim/plugins/indent-blankline.lua
@@ -1,0 +1,44 @@
+return {
+  "lukas-reineke/indent-blankline.nvim",
+  main = "ibl",
+  event = "User AstroFile",
+  cmd = {
+    "IBLEnable",
+    "IBLDisable",
+    "IBLToggle",
+    "IBLEnableScope",
+    "IBLDisableScope",
+    "IBLToggleScope",
+  },
+  opts = {
+    indent = { char = "▏" },
+    scope = {
+      enabled = true,
+      show_start = false,
+      show_end = false,
+    },
+    exclude = {
+      buftypes = {
+        "nofile",
+        "prompt",
+        "quickfix",
+        "terminal",
+      },
+      filetypes = {
+        "aerial",
+        "alpha",
+        "dashboard",
+        "help",
+        "lazy",
+        "mason",
+        "neo-tree",
+        "NvimTree",
+        "neogitstatus",
+        "notify",
+        "startify",
+        "toggleterm",
+        "Trouble",
+      },
+    },
+  },
+}
