--- lua/deltavim/plugins/indent-blankline/opts.lua
@@ -1,31 +1,0 @@
-return {
-  indent = { char = "▏" },
-  scope = {
-    enabled = true,
-    show_start = false,
-    show_end = false,
-  },
-  exclude = {
-    buftypes = {
-      "nofile",
-      "prompt",
-      "quickfix",
-      "terminal",
-    },
-    filetypes = {
-      "aerial",
-      "alpha",
-      "dashboard",
-      "help",
-      "lazy",
-      "mason",
-      "neo-tree",
-      "NvimTree",
-      "neogitstatus",
-      "notify",
-      "startify",
-      "toggleterm",
-      "Trouble",
-    },
-  },
-}
