+++ lua/deltavim/plugins/mini-bufremove.lua
@@ -1,0 +1,1 @@
+return { "echasnovski/mini.bufremove", lazy = true }
