--- lua/deltavim/plugins/notify/initialize.lua
@@ -1,1 +1,0 @@
-return function() require("astrocore").load_plugin_with_func("nvim-notify", vim, "notify") end
