--- lua/deltavim/plugins/notify/setup.lua
@@ -1,5 +1,0 @@
-return function(_, opts)
-  local notify = require "notify"
-  notify.setup(opts)
-  require("deltavim.notify").setup(notify)
-end
