+++ lua/deltavim/plugins/todo-comments.lua
@@ -1,0 +1,8 @@
+return {
+  "folke/todo-comments.nvim",
+  event = "User AstroFile",
+  cmd = { "TodoTrouble", "TodoTelescope", "TodoLocList", "TodoQuickFix" },
+  opts = {
+    highlight = { multiline = true },
+  },
+}
