--- lua/deltavim/plugins/ufo/opts.lua
@@ -1,30 +1,0 @@
-local function handleFallbackException(bufnr, err, providerName)
-  if type(err) == "string" and err:match "UfoFallbackException" then
-    return require("ufo").getFolds(bufnr, providerName)
-  else
-    return require("promise").reject(err)
-  end
-end
-
-return {
-  preview = {
-    win_config = {
-      border = require("deltavim").get_border("popup_border", "Folded"),
-    },
-    mappings = {
-      scrollB = "<C-b>",
-      scrollF = "<C-f>",
-      scrollU = "<C-u>",
-      scrollD = "<C-d>",
-    },
-  },
-  provider_selector = function(_, filetype, buftype)
-    return (filetype == "" or buftype == "nofile") and "indent" -- only use indent until a file is opened
-      or function(bufnr)
-        return require("ufo")
-          .getFolds(bufnr, "lsp")
-          :catch(function(err) return handleFallbackException(bufnr, err, "treesitter") end)
-          :catch(function(err) return handleFallbackException(bufnr, err, "indent") end)
-      end
-  end,
-}
