--- lua/deltavim/plugins/which-key/opts.lua
@@ -1,8 +1,0 @@
-return {
-  icons = {
-    group = vim.g.icons_enabled ~= false and "" or "+",
-    separator = "-",
-  },
-  disable = { filetypes = { "TelescopePrompt" } },
-  window = { border = require("deltavim").get_border "popup_border" },
-}
